// src/constants/enums.ts
export enum Chains {
  ETHEREUM = "ethereum",
  BNBCHAIN = "bnb",
  POLYGON = "polygon",
  AVALANCHE = "avalanche",
  ARBITRUM = "arbitrum",
  BASE = "base",
  SCROLL = "scroll",
  ZKSYNC = "zksync",
  SEI = "sei",
  SOLANA = "solana",
  TON = "ton",
}

export enum ChainsNativeToken {
  ethereum = "ETH",
  bnb = "BNB",
  polygon = "MATIC",
  avalanche = "AVAX",
  arbitrum = "ETH",
  base = "ETH",
  scroll = "ETH",
  zksync = "ETH",
  sei = "SEI",
  solana = "SOL",
  ton = "TON",
}

export enum Tokens {
  USDT = "usdt",
  USDC = "usdc",
  PYUSD = "pyusd",
}

export enum SubscriptionPayCycle {
  Daily = "/daily",
  Weekly = "/weekly",
  Monthly = "/monthly",
  Yearly = "/yearly",
}